Update scip.rs
Change the scip.rs to accept an extra parameter in the function write, so the user can decide whether they want the name of the variable or the default x_{varnumber} in the written file. I am not the best coder so I appreciate if this pull request is reviewed.