upd: change `col` method of `Variable` struct.
Now, we do not check if the variable is in the LP anymore. Instead, we simply check if the variable has a column associated with it. This is because newly generated columns will not be in the LP, however they are columns and can have reduced cost associated with them.